Fix estaciones mapping to real schema columns

+++ src/main/java/com/dv/agro_web/controllers/EstacionesController.java
@@ -31,7 +31,7 @@ public class EstacionesController {
     }
 
     @GetMapping("/estaciones/{id}/configurar")
-    public String configurarEstacion(@PathVariable Integer id, Model model) {
+    public String configurarEstacion(@PathVariable Long id, Model model) {
         model.addAttribute("estacionId", id);
         return "estaciones-configurar";
     }
