+++ src/main/java/com/dv/agro_web/repositorios/EstacionRepository.java
@@ -3,5 +3,5 @@ package com.dv.agro_web.repositorios;
 import com.dv.agro_web.entidades.Estacion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EstacionRepository extends JpaRepository<Estacion, Integer> {
+public interface EstacionRepository extends JpaRepository<Estacion, Long> {
 }
